Statistical: compute the n-squared denominators in double -- Int32 overflows above 46,340.
Variance and Covariance form their combined denominator as an Int32 product (n * n,
and n * (n - 1) for the sample forms). That product overflows for any count above
46,340, silently producing wrong variances; at exactly 2^20 entries it wraps to zero
and the term becomes an infinity, so correlation over a 1,048,576-row set returns NaN.

Computing the product in double fixes every large count while changing nothing at the
sizes that already worked: wherever the Int32 product did not overflow, the same value
is exactly representable in double, so those results are unchanged bit for bit.

The new test pins the failure sizes with exact expectations (alternating 0/1 values
keep every intermediate sum a power of two): population variance 0.25 and
self-correlation 1 at 2^20 entries, plus the smallest overflowing count, 46,341.
Verified against the unfixed code: the test fails there and passes with the fix.

+++ source/Statistical.cs
@@ -16,14 +16,18 @@ public static class Statistical
 		// (sum2 - sum * sum / n) / (n - 1) // Sample
 
 		// Reduce the amount of division in order to reduce the amount of double precision error.
+		// The n·n product must be computed in double: as Int32 it overflows for any n above
+		// 46,340 (and at n = 2^20 wraps to exactly zero, turning the term into an infinity).
+		// For every n it does not overflow at, the product is an exactly representable double,
+		// so results at those sizes are unchanged bit for bit.
 		if (!sample)
-			return sum2 / n - sum * sum / (n * n);
+			return sum2 / n - sum * sum / ((double)n * n);
 
 		if (n == 1)
 			return double.NaN; // Avoid divide by zero.
 
 		var n1 = n - 1;
-		var n2 = n * n1;
+		var n2 = (double)n * n1;
 
 		return sum2 / n1 - sum * sum / n2;
 	}
@@ -222,11 +226,13 @@ public static class Statistical
 
 	static double Covariance(int n, double sumA, double sumB, double prod, bool sample)
 	{
+		// The n·n product in double for the same reason as Variance above: Int32 overflows for
+		// any n above 46,340, and identical values everywhere it does not.
 		if (!sample)
-			return prod / n - sumA * sumB / (n * n);
+			return prod / n - sumA * sumB / ((double)n * n);
 
 		var n2 = n - 1;
-		return prod / n2 - sumA * sumB / (n * n2);
+		return prod / n2 - sumA * sumB / ((double)n * n2);
 	}
 
 	/// <summary>
